re-enable legacy blend syntax
